# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Blickling Fieldbook API
servers:
  - url: /api
    description: Base API path
...
security:
  - cookieAuth: []
paths:
...
  /observations:
    get:
      operationId: listObservations
      tags: [observations]
      summary: List observations with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: namedLocationId
          in: query
          schema:
            type: integer
        - name: assignedUserId
          in: query
          schema:
            type: integer
        - name: safetyIssue
          in: query
          schema:
            type: boolean
        - name: publicAccessAffected
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated observation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservationList"

    post:
      operationId: createObservation
      tags: [observations]
      summary: Create a new observation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObservationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Observation"
  /observations/map:
    get:
      operationId: getObservationMapData
      tags: [observations]
      summary: Get lightweight observation data for map markers
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: namedLocationId
          in: query
          schema:
            type: integer
        - name: safetyIssue
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Map marker data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ObservationMarker"
  /observations/{id}:
    get:
      operationId: getObservation
      tags: [observations]
      summary: Get full observation detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Observation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservationDetail"

    patch:
      operationId: updateObservation
      tags: [observations]
      summary: Update an observation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObservationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Observation"

    delete:
      operationId: deleteObservation
      tags: [observations]
      summary: Delete an observation (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /observations/{id}/status:
    patch:
      operationId: updateObservationStatus
      tags: [observations]
      summary: Update observation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Observation"

  # Actions
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: blickling.sid
  schemas:
...
    Observation:
      type: object
      required: [id, referenceNumber, title, priority, status, createdAt, reportedByUserId, categoryId]
      properties:
        id:
          type: integer
        referenceNumber:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        categoryColour:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        status:
          type: string
          enum: [draft, submitted, under_review, action_required, monitoring, resolved, closed, cancelled]
        observedAt:
          type: string
        reportedByUserId:
          type: integer
        reportedByName:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        gpsAccuracyMetres:
          type: ["number", "null"]
        namedLocationId:
          type: ["integer", "null"]
        namedLocationName:
          type: ["string", "null"]
        safetyIssue:
          type: boolean
        publicAccessAffected:
          type: boolean
        machineryRequired:
          type: boolean
        specialistRequired:
          type: boolean
        followUpRequired:
          type: boolean
        actionCount:
          type: integer
        propertyId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ObservationDetail:
      type: object
      required: [id, referenceNumber, title, priority, status, createdAt, reportedByUserId, categoryId]
      properties:
        id:
          type: integer
        referenceNumber:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        categoryColour:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        status:
          type: string
          enum: [draft, submitted, under_review, action_required, monitoring, resolved, closed, cancelled]
        observedAt:
          type: string
        reportedByUserId:
          type: integer
        reportedByName:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        gpsAccuracyMetres:
          type: ["number", "null"]
        namedLocationId:
          type: ["integer", "null"]
        namedLocationName:
          type: ["string", "null"]
        safetyIssue:
          type: boolean
        publicAccessAffected:
          type: boolean
        machineryRequired:
          type: boolean
        specialistRequired:
          type: boolean
        followUpRequired:
          type: boolean
        propertyId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        actions:
          type: array
          items:
            $ref: "#/components/schemas/Action"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        auditEvents:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
    ObservationList:
      type: object
      required: [observations, total, page, limit]
      properties:
        observations:
          type: array
          items:
            $ref: "#/components/schemas/Observation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ObservationMarker:
      type: object
      required: [id, title, priority, status, latitude, longitude]
      properties:
        id:
          type: integer
        title:
          type: string
        referenceNumber:
          type: string
        priority:
          type: string
        status:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        categoryName:
          type: ["string", "null"]
        categoryColour:
          type: ["string", "null"]
        namedLocationName:
          type: ["string", "null"]
        safetyIssue:
          type: boolean
    ObservationInput:
      type: object
      required: [title, categoryId, priority, observedAt]
      properties:
        title:
          type: string
          maxLength: 100
        description:
          type: string
        categoryId:
          type: integer
        priority:
          type: string
          enum: [low, normal, high, urgent]
        status:
          type: string
          enum: [draft, submitted]
        observedAt:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        gpsAccuracyMetres:
          type: number
        namedLocationId:
          type: integer
        safetyIssue:
          type: boolean
        publicAccessAffected:
          type: boolean
        machineryRequired:
          type: boolean
        specialistRequired:
          type: boolean
        followUpRequired:
          type: boolean
        createdOffline:
          type: boolean
        offlineId:
          type: string
    ObservationUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        priority:
          type: string
          enum: [low, normal, high, urgent]
        observedAt:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        gpsAccuracyMetres:
          type: ["number", "null"]
        namedLocationId:
          type: ["integer", "null"]
        safetyIssue:
          type: boolean
        publicAccessAffected:
          type: boolean
        machineryRequired:
          type: boolean
        specialistRequired:
          type: boolean
        followUpRequired:
          type: boolean
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        reason:
          type: string

    # Actions
    Action:
      type: object
      required: [id, referenceNumber, title, priority, status, createdAt, createdByUserId]
      properties:
        id:
          type: integer
        referenceNumber:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        observationId:
          type: ["integer", "null"]
        observationTitle:
          type: ["string", "null"]
        observationRef:
          type: ["string", "null"]
        assignedToUserId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        createdByUserId:
          type: integer
        priority:
          type: string
          enum: [low, normal, high, urgent]
        status:
          type: string
          enum: [not_started, planned, in_progress, waiting, completed, cancelled]
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        completionNote:
          type: ["string", "null"]
        namedLocationName:
          type: ["string", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        equipmentRequired:
          type: boolean
        contractorRequired:
          type: boolean
        propertyId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Note:
      type: object
      required: [id, body, createdByUserId, createdAt]
      properties:
        id:
          type: integer
        body:
          type: string
        observationId:
          type: ["integer", "null"]
        actionId:
          type: ["integer", "null"]
        offlineId:
          type: ["string", "null"]
        createdByUserId:
          type: integer
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AuditEvent:
      type: object
      required: [id, eventType, createdAt]
      properties:
        id:
          type: integer
        eventType:
          type: string
        fieldName:
          type: ["string", "null"]
        previousValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        createdAt:
          type: string

    # Dashboard
...